game updates, improvements
