+++ src/modules/game/game.card.tsx
