+++ src/pages/game.page.tsx
@@ -1,0 +1,35 @@
+import { useParams } from 'react-router-dom';
+import { useEffect } from 'react';
+import { useInitData } from '@vkruglikov/react-telegram-web-app';
+import { useAuth } from '@/shared/hooks/useAuth';
+import { useLobbyStore } from '@/shared/stores/lobby.store';
+import { userEvents } from '@/shared/events/app-events/user.event';
+
+import { useRoutes } from '@/shared/hooks/useRoutes';
+import { GameComponent } from '@/components/ui/game';
+
+const GamePage = () => {
+  const { setLobbyId, lobbyId } = useLobbyStore();
+  const { id } = useParams<{ id: string }>(); //lobbyId
+  const { user, authenticated, loginUser, ready } = useAuth();
+  const [initDataUnsafe] = useInitData();
+  const routes = useRoutes();
+
+  useEffect(() => {
+    if (user === undefined && initDataUnsafe?.user && ready) {
+      loginUser({
+        name: initDataUnsafe.user.first_name,
+        avatar: '',
+        telegram: initDataUnsafe.user.id,
+      });
+    }
+    if (id && authenticated && lobbyId == undefined) {
+      setLobbyId(id);
+      userEvents.joinLobby(id, user?.id);
+    }
+  }, [id, user, ready]);
+
+  return <GameComponent />;
+};
+
+export default GamePage;
